+++ NEWS.md
@@ -12,3 +12,6 @@
 - Added function for Lin's Concordance Correlation Coefficient.
 - Added function SiGN eval (incomplete)
   - To Do: Add $\phi$, log-likelihood, AIC, and BIC, for Beta Distribution error model. Throw warning at low N.
+
+## 2025-07-28
+- Added function to compute (negative) log-Likelihood for a Beta error model